pulled apart the translation functions into another file
